+++ CHANGELOG.md
@@ -70,3 +70,16 @@ any extra access details
 
 - Updated default user agent string so that bogus data will not be collected during testing
 - Updated service account bindings to include ProjectId
+
+## [3.1.0] - 2017-02-07
+
+- Updated vendored packages so that custom UserAgent string gets propagated for storage provision calls
+- Added Bigtable as a service
+
+## [3.1.1] - 2017-02-09
+
+- fixed a bug in reading in custom plans
+
+## [3.1.2] - 2017-02-12
+
+- fixed a bug where supplying a custom name for Bigtable would cause an error